# openapi/workos-subset.yaml
openapi: 3.0.3
info:
  title: WorkOS Subset
  version: 0.1.0
servers:
  - url: https://api.workos.com
paths:
...
  /user_management/authenticate:
    post:
      operationId: authenticate
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/DeviceAuthenticateRequest"
                - $ref: "#/components/schemas/RefreshAuthenticateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticateResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    DeviceAuthenticateRequest:
      type: object
      properties:
        client_id:
          type: string
        grant_type:
          type: string
          const: urn:ietf:params:oauth:grant-type:device_code
        device_code:
          type: string
      required:
        - client_id
        - grant_type
        - device_code
    RefreshAuthenticateRequest:
      type: object
      properties:
        client_id:
          type: string
        client_secret:
          type: string
        grant_type:
          type: string
          const: refresh_token
        refresh_token:
          type: string
      required:
        - client_id
        - client_secret
        - grant_type
        - refresh_token
    AuthenticateResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        organization_id:
          type: string
          nullable: true
        access_token:
          type: string
        refresh_token:
          type: string
      required:
        - user
        - access_token
        - refresh_token
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        profile_picture_url:
          type: string
          nullable: true
      required:
        - id
        - email
        - first_name
        - last_name
        - profile_picture_url
...